+++ src/main/java/com/odeyalo/analog/auth/dto/request/LoginUserDTO.java
@@ -1,16 +1,11 @@
-package com.odeyalo.analog.auth.dto;
+package com.odeyalo.analog.auth.dto.request;
 
-public class RegisterUserDTO {
-
+public class LoginUserDTO {
     private String nickname;
-    private String email;
     private String password;
 
-    public RegisterUserDTO() {}
-
-    public RegisterUserDTO(String nickname, String email, String password) {
+    public LoginUserDTO(String nickname, String password) {
         this.nickname = nickname;
-        this.email = email;
         this.password = password;
     }
 
@@ -20,14 +15,6 @@ public class RegisterUserDTO {
 
     public void setNickname(String nickname) {
         this.nickname = nickname;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getPassword() {
