+++ src/main/java/com/odeyalo/analog/auth/dto/request/NewPasswordDTO.java
@@ -1,4 +1,4 @@
-package com.odeyalo.analog.auth.dto;
+package com.odeyalo.analog.auth.dto.request;
 
 public class NewPasswordDTO {
     private String password;
