+++ src/main/java/com/odeyalo/analog/auth/dto/request/PhoneNumberMethodPasswordRecoveryDTO.java
@@ -1,4 +1,4 @@
-package com.odeyalo.analog.auth.dto;
+package com.odeyalo.analog.auth.dto.request;
 
 public class PhoneNumberMethodPasswordRecoveryDTO {
     private String phoneNumber;
